api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arize-phoenix==13.15.0
arize-phoenix-client==2.1.0
arize-phoenix-evals==2.11.0
arize-phoenix-otel==0.15.0
attrs==25.4.0
Authlib==1.6.9
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cross-web==0.4.1
cryptography==46.0.5
dataclasses-json==0.6.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
graphql-core==3.2.8
greenlet==3.3.2
grpc-interceptor==0.15.4
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-huggingface==1.2.1
langchain-ollama==1.0.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
ldap3==2.9.1
lia-web==0.3.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.3
ollama==0.6.1
openinference-instrumentation==0.1.46
openinference-instrumentation-langchain==0.1.61
openinference-instrumentation-openai==0.1.42
openinference-semantic-conventions==0.1.28
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
portalocker==3.2.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
sqlean.py==3.49.1
starlette==0.52.1
strawberry-graphql==0.287.3
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
